requirements.txt
absl-py==0.7.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
cloudpickle==0.8.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dask==1.1.4
decorator==4.4.0
defer==1.0.6
distro-info==0.18
Django==2.1.7
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
httplib2==0.9.2
idna==2.6
ipython-genutils==0.2.0
jupyter-core==4.4.0
kazam==1.4.5
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.3
mock==2.0.0
mysql-connector==2.1.6
netifaces==0.10.4
networkx==2.2
numpy==1.16.1
oauth==1.0.1
olefile==0.45.1
pandas==0.24.1
pbr==5.1.3
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.7.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.1
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.0.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.14.2
scipy==1.2.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
toolz==0.9.0
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==0.15.0
xkit==0.0.0
zope.interface==4.3.2
